+++ logger/logger.py
@@ -1,24 +1,22 @@
 __G__ = "(G)bd249ce4"
 
-from logging import DEBUG, ERROR, Formatter, StreamHandler, WARNING, getLogger, handlers
+from logging import DEBUG, ERROR, Handler, WARNING, getLogger
 from os import path
 from _thread import interrupt_main
 from sys import stdout,stderr
 from threading import Timer
 from datetime import datetime
 from ..settings import json_settings
-from ..connections.mongodbconn import add_item_fs
+from ..connections.mongodbconn import add_item_fs,add_item
 from tempfile import gettempdir
 
-logterminal,logfile,dynamic,verbose_flag,verbose_timeout = None,None, None, None, None
+logterminal,dynamic,verbose_flag,verbose_timeout = None,None, None, None
 
 if dynamic == None:dynamic = getLogger("qbanalyzerdynamic")
 if logterminal == None:logterminal = getLogger("qbanalyzerlogterminal")
-if logfile == None:logfile = getLogger("qbanalyzerlogfile")
 if verbose_flag == None:verbose_flag = False
 if verbose_timeout == None: verbose_timeout = json_settings["function_timeout"]
 
-logterminalgerpath = None
 lock = False
 chain = []
 
@@ -33,20 +31,56 @@ class colors:
     Cyan="\033[36m"
     White="\033[37m"
 
+green_x = '{}{}{}'.format(colors.Green,"X",colors.Restore)
+exclamation_mark = '{}{}{}'.format(colors.Yellow,">",colors.Restore)
+red_mark = '{}{}{}'.format(colors.Red,">",colors.Restore)
+yellow_hashtag = '{}{}{}'.format(colors.Yellow,"#",colors.Restore)
+
+class Unbuffered:
+   def __init__(self, stream):
+       self.stream = stream
+
+   def write(self, data):
+       self.stream.write(data)
+       self.stream.flush()
+       te.write(data)
+
+class CustomHandler(Handler):
+    def __init__(self,file):
+        Handler.__init__(self)
+        self.logsfile = open(file,'w')
+
+    def emit(self, record):
+        print("{} {} {}".format(record.msg[0], record.msg[2], record.msg[1]))
+        stdout.flush()
+        self.logsfile.write("{} {} {}\n".format(record.msg[0],record.msg[2],record.msg[1]))
+        self.logsfile.flush()
+        add_item("analyzer","alllogs",{'time':record.msg[0],'message': record.msg[1]})
+
+class TaskHandler(Handler):
+    def __init__(self,task):
+        Handler.__init__(self)
+        self.logsfile = open(path.join(gettempdir(),task),'w')
+        self.task = task
+
+    def emit(self, record):
+        self.logsfile.write("{} {}\n".format(record.msg[0],record.msg[1]))
+        self.logsfile.flush()
+        add_item("analyzer","tasklogs",{'time':record.msg[0],"task":self.task,'message': record.msg[1]})
+
 def setup_task_logger(task):
-    format = Formatter('%(asctime)s %(message)s',datefmt='%Y-%m-%d %H:%M:%S')
+    log_string("Setting up task {} logger".format(task),"Yellow")
+    dynamic.handlers.clear()
     dynamic.setLevel(DEBUG)
-    fhandler = handlers.RotatingFileHandler(path.join(logterminalgerpath,task), maxBytes=(100*1024*1024), backupCount=3)
-    fhandler.setFormatter(format)
-    dynamic.addHandler(fhandler)
+    dynamic.addHandler(TaskHandler(task))
     dynamic.disabled = False
 
 def cancel_task_logger(task):
+    log_string("Closing up task {} logger".format(task),"Yellow")
     dynamic.disabled = True
     logs = ""
-    with open(path.join(logterminalgerpath,task),"rb") as f:
+    with open(path.join(gettempdir(),task),"rb") as f:
         logs = f.read()
-
     if len(logs) > 0:
         _id = add_item_fs("webinterface","logs",logs,"log",None,task,"text/plain",datetime.now())
         if _id:
@@ -60,23 +94,19 @@ def log_string(_str,color):
     '''
     output str with color and symbol (they are all as info)
     '''
-    stdout.flush()
+    ctime = datetime.now()
     if color == "Green":
-        logterminal.info('{}{}{} {}'.format(colors.Green,"X",colors.Restore,_str))
-        dynamic.info('{} {}'.format("X",_str))
-        logfile.info('{} {}'.format("X",_str))
+        logterminal.info([ctime,_str,green_x])
+        dynamic.info([ctime,_str,"X"])
     elif color == "Yellow":
-        logterminal.info('{}{}{} {}'.format(colors.Yellow,">",colors.Restore,_str))
-        dynamic.info('{} {}'.format(">",_str))
-        logfile.info('{} {}'.format(">",_str))
+        logterminal.info([ctime,_str,exclamation_mark])
+        dynamic.info([ctime,_str,"!"])
     elif color == "Red":
-        logterminal.info('{}{}{} {}'.format(colors.Red,"!",colors.Restore,_str))
-        dynamic.info('{} {}'.format("!",_str))
-        logfile.info('{} {}'.format("!",_str))
-    elif color == "Yellow_#":
-        logterminal.info('{}{}{} {}'.format(colors.Yellow,"#",colors.Restore,_str))
-        dynamic.info('{} {}'.format("#",_str))
-        logfile.info('{} {}'.format("#",_str))
+        logterminal.info([ctime,_str,red_mark])
+        dynamic.info([ctime,_str,"!"])
+    elif color == "Yellow":
+        logterminal.info([ctime,_str,yellow_hashtag])
+        dynamic.info([ctime,_str,"#"])
 
 def verbose(OnOff=False,Verb=False,timeout=30,str=None,extra=None):
     '''
@@ -88,7 +118,6 @@ def verbose(OnOff=False,Verb=False,timeout=30,str=None,extra=None):
         interrupt_main()
         lock = fn
         log_string("{} > {}s.. Timeout".format(fn,timeout), "Red")
-        #print(">>>>>>>>>>>>>>>>>>" + lock)
 
     def decorator(func):
         def wrapper(*args, **kwargs):
@@ -100,8 +129,6 @@ def verbose(OnOff=False,Verb=False,timeout=30,str=None,extra=None):
             timer = None
             ret = None
             chain.append(function_name)
-            #print(chain)
-            #print(tenumerate())
             try:
                 if Verb:
                     log_string("Function '{0}', parameters : {1} and {2}".format(func.__name__, args, kwargs))
@@ -118,8 +145,7 @@ def verbose(OnOff=False,Verb=False,timeout=30,str=None,extra=None):
                 pass
             except Exception as e:
                 #print(e)
-                log_string("{}.{} Failed".format(func.__module__, func.__name__), "Red")
-                logfile.info("{}.{} Failed -> {}".format(func.__module__, func.__name__,e))
+                log_string("{}.{} Failed -> {}".format(func.__module__, func.__name__,e),"Red")
             finally:
                 if timer != None:
                     timer.cancel()
@@ -132,8 +158,6 @@ def verbose(OnOff=False,Verb=False,timeout=30,str=None,extra=None):
     return decorator
 
 def setup_logger():
-    global logterminalgerpath
-    logterminalgerpath = gettempdir()
     getLogger("scapy.runtime").setLevel(ERROR)
     getLogger("requests").setLevel(WARNING)
     getLogger("urllib3").setLevel(WARNING)
@@ -141,11 +165,4 @@ def setup_logger():
     getLogger("PIL").setLevel(WARNING)
     getLogger("chardet").setLevel(WARNING)
     logterminal.setLevel(DEBUG)
-    format = Formatter('%(asctime)s %(message)s',datefmt='%Y-%m-%d %H:%M:%S')
-    shandler = StreamHandler(stdout)
-    shandler.setFormatter(format)
-    logterminal.addHandler(shandler)
-    logfile.setLevel(DEBUG)
-    fhandler = handlers.RotatingFileHandler(path.join(logterminalgerpath,"alllogterminals"), maxBytes=(100*1024*1024), backupCount=3)
-    fhandler.setFormatter(format)
-    logfile.addHandler(fhandler)
+    logterminal.addHandler(CustomHandler(path.join(gettempdir(),"alllogs")))