+++ modules/archive.py
@@ -1,3 +1,5 @@
+__G__ = "(G)bd249ce4"
+
 from ..logger.logger import verbose, verbose_flag, verbose_timeout
 from os import path, walk
 from subprocess import PIPE,Popen