+++ services/online/malshare.py
@@ -1,3 +1,5 @@
+__G__ = "(G)bd249ce4"
+
 from shutil import copyfileobj
 from requests import get
 from requests.packages.urllib3 import disable_warnings