+++ services/online/metadefender.py
@@ -1,3 +1,5 @@
+__G__ = "(G)bd249ce4"
+
 from requests import get
 from ...logger.logger import verbose, verbose_flag, verbose_timeout
 from json import load,dumps