+++ start.py
@@ -1,8 +1,7 @@
 from mics.certmaker import create_dummy_certificate
 from os import path, mkdir
-from uuid import uuid4
-from sys import stdout,argv
-from os import path,listdir,environ
+from sys import argv
+from os import environ, path
 
 if __name__ == '__main__':
     if len(argv) == 2: